feat(14j): move "Add section" to below last section in canvas
Eric's feedback: the Add section button at the top of the page is
disconnected from the place where the section actually lands. Move it
to where you'd expect it — at the end of the section list.

- Page header right side now contains only the version pill
- Removed the inline section-add panel that sat under the header
- Canvas renders an "+ Add section" button below the last sortable
  section card. Clicking it swaps to an inline title input + Add /
  Cancel buttons, in the same canvas position.
- Local state (showAddSection, newSectionTitle) and the addSection
  function moved from page.tsx to canvas.tsx since Canvas already
  has setConfig.
